rearranged dependencies among Prelude.notnull, Prelude.isnull. and IsNull extension
