+++ LanguageExt.Core/Prelude/Prelude.cs
@@ -330,7 +330,7 @@ namespace LanguageExt
         [Pure]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool isnull<T>(T value) =>
-            NullChecker<T>.IsNull(value);
+            value.IsNull();
 
         /// <summary>
         /// Returns true if the value is not null, and does so without
@@ -352,7 +352,7 @@ namespace LanguageExt
         [Pure]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool notnull<T>(T value) =>
-            !NullChecker<T>.IsNull(value);
+            !isnull(value);
 
         /// <summary>
         /// Convert a value to string
@@ -360,25 +360,5 @@ namespace LanguageExt
         [Pure]
         public static string toString<T>(T value) =>
             value?.ToString();
-
-        static class NullChecker<T>
-        {
-            static readonly bool IsReferenceType;
-            static readonly bool IsNullable;
-            static readonly EqualityComparer<T> DefaultEqualityComparer;
-
-            static NullChecker()
-            {
-                IsNullable = Nullable.GetUnderlyingType(typeof(T)) != null;
-                IsReferenceType = !typeof(T).GetTypeInfo().IsValueType;
-                DefaultEqualityComparer = EqualityComparer<T>.Default;
-            }
-
-            [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            public static bool IsNull(T value) =>
-                IsNullable
-                    ? value.Equals(default(T))
-                    : IsReferenceType && DefaultEqualityComparer.Equals(value, default(T));
-        }
    }
 }
